+++ server/src/main/java/org/guohai/javasqlweb/beans/WebAuthnRequestType.java
@@ -1,0 +1,9 @@
+package org.guohai.javasqlweb.beans;
+
+/**
+ * WebAuthn 请求类型
+ */
+public enum WebAuthnRequestType {
+    ASSERTION,
+    REGISTRATION
+}